requirements.txt
# Core Framework - Minimal Setup (No Build Required)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Financial Data
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2

# Configuration
python-dotenv==1.0.0

# Payment Processing (optional)
stripe==7.7.0
